requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
apscheduler
requests
selenium
selenium-wire
